modify database scheme

+++ src/controllers/orders.controller.ts
@@ -1,9 +1,10 @@
-import { NextFunction, Request, Response } from 'express';
+import { NextFunction, Response } from 'express';
 import { Container } from 'typedi';
 import { CreateOrderDto } from '@/dtos/orders.dto';
 import { Order } from '@interfaces/orders.interface';
 import { OrderService } from '@/services/orders.service';
 import { RequestWithUser, Role } from '@/interfaces/auth.interface';
+import { HttpException } from '@/exceptions/httpException';
 
 export class OrderController {
   public order = Container.get(OrderService);
@@ -12,8 +13,9 @@ export class OrderController {
     try {
       const { user } = req;
       let findAllOrdersData: Order[];
+      const acceptedRoles = [Role.ADMIN, Role.DELIVERER];
 
-      if (user.role === Role.ADMIN) {
+      if (acceptedRoles.includes(user.getDataValue('role'))) {
         findAllOrdersData = await this.order.findAllOrders();
       } else {
         findAllOrdersData = await this.order.findAllOrdersByUserId(user.id);
@@ -25,9 +27,14 @@ export class OrderController {
     }
   };
 
-  public getOrderById = async (req: Request, res: Response, next: NextFunction) => {
+  public getOrderById = async (req: RequestWithUser, res: Response, next: NextFunction) => {
     try {
       const orderId = Number(req.params.id);
+
+      if (req.user.role === Role.CUSTOMER && !this.order.isOrderBelongsToUser(orderId, req.user.id)) {
+        throw new HttpException(404, "Order doesn't exist");
+      }
+
       const findOneOrderData: Order = await this.order.findOrderById(orderId);
 
       res.status(200).json({ data: findOneOrderData, message: 'findOne' });
@@ -47,21 +54,30 @@ export class OrderController {
     }
   };
 
-  public updateOrder = async (req: Request, res: Response, next: NextFunction) => {
+  public updateOrder = async (req: RequestWithUser, res: Response, next: NextFunction) => {
     try {
       const orderId = Number(req.params.id);
       const orderData: CreateOrderDto = req.body;
-      const updateorderData: Order = await this.order.updateOrder(orderId, orderData);
+      const updateOrderData: Order = await this.order.updateOrder(orderId, orderData);
 
-      res.status(200).json({ data: updateorderData, message: 'updated' });
+      if (req.user.role === Role.CUSTOMER && !this.order.isOrderBelongsToUser(orderId, req.user.id)) {
+        throw new HttpException(404, "Order doesn't exist");
+      }
+
+      res.status(200).json({ data: updateOrderData, message: 'updated' });
     } catch (error) {
       next(error);
     }
   };
 
-  public deleteOrder = async (req: Request, res: Response, next: NextFunction) => {
+  public deleteOrder = async (req: RequestWithUser, res: Response, next: NextFunction) => {
     try {
       const orderId = Number(req.params.id);
+
+      if (req.user.role === Role.CUSTOMER && !this.order.isOrderBelongsToUser(orderId, req.user.id)) {
+        throw new HttpException(404, "Order doesn't exist");
+      }
+
       const deleteOrderData: Order = await this.order.deleteOrder(orderId);
 
       res.status(200).json({ data: deleteOrderData, message: 'deleted' });
